tools/requirements.txt
requests==2.32.3
feedparser==6.0.11
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
pandas>=2.0,<2.1
PyYAML==6.0.2
matplotlib>=3.7,<3.9
